requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
_openmp_mutex=4.5=2_gnu
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
anaconda-anon-usage=0.5.0=py312hfc23b7f_100
anaconda_powershell_prompt=1.1.0=haa95532_0
anaconda_prompt=1.1.0=haa95532_0
archspec=0.2.3=pyhd3eb1b0_0
attrs=25.3.0=pypi_0
boltons=23.0.0=py312haa95532_0
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.1.31=h56e8100_0
cairo=1.18.4=h5782bbf_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py312h827c3e9_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cloudpickle=3.1.1=pypi_0
colorama=0.4.6=py312haa95532_0
conda=24.11.3=py312h2e8e312_0
conda-anaconda-telemetry=0.1.1=py312haa95532_0
conda-content-trust=0.2.0=py312haa95532_1
conda-libmamba-solver=24.9.0=pyhd3eb1b0_0
conda-package-handling=2.4.0=py312haa95532_0
conda-package-streaming=0.11.0=py312haa95532_0
contourpy=1.3.1=pypi_0
cryptography=43.0.3=py312hbd6ee87_1
cycler=0.12.1=pypi_0
distro=1.9.0=py312haa95532_0
expat=2.6.4=h8ddb27b_0
farama-notifications=0.0.4=pypi_0
filelock=3.18.0=pypi_0
fmt=9.1.0=h6d14046_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h765892d_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.57.0=pypi_0
freetype=2.13.3=h0b5ce68_0
fribidi=1.0.10=h8d14728_0
frozendict=2.4.2=py312haa95532_0
frozenlist=1.6.0=pypi_0
fsspec=2025.3.2=pypi_0
getopt-win32=0.1=hcfcfb64_1
graphite2=1.3.13=h63175ca_1003
graphviz=12.2.1=hf40819d_1
gts=0.7.6=h6b5321d_4
gymnasium=0.29.1=pypi_0
harfbuzz=11.0.0=h9e37d49_0
icu=75.1=he0c23c2_0
idna=3.7=py312haa95532_0
importlib-metadata=8.7.0=pypi_0
jinja2=3.1.6=pypi_0
jsonpatch=1.33=py312haa95532_1
jsonpointer=2.1=pyhd3eb1b0_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=hdf4eb48_0
lerc=4.0.0=h63175ca_0
libarchive=3.7.7=h979ed78_3
libcurl=8.13.0=h88aaa65_0
libdeflate=1.23=h9062f6e_0
libexpat=2.6.4=he0c23c2_0
libffi=3.4.4=hd77b12b_1
libgcc=14.2.0=h1383e82_2
libgd=2.3.3=h7208af6_11
libglib=2.84.0=h7025463_0
libgomp=14.2.0=h1383e82_2
libiconv=1.18=h135ad9c_1
libintl=0.22.5=h5728263_3
libjpeg-turbo=3.0.0=hcfcfb64_1
liblzma=5.8.1=h2466b09_0
libmamba=1.5.11=h99b1521_1
libmambapy=1.5.11=py312hcd6fe79_1
libpng=1.6.47=had7236b_0
libsolv=0.7.32=hbb528cf_1
libsqlite=3.49.1=h67fdade_2
libssh2=1.11.1=he619c9f_0
libtiff=4.7.0=h797046b_3
libwebp-base=1.5.0=h3b0e114_0
libwinpthread=12.0.0.r4.gg4f2fc60ca=h57928b3_9
libxcb=1.17.0=h0e4246c_0
libxml2=2.13.7=h442d1da_1
libzlib=1.3.1=h2466b09_2
lz4-c=1.10.0=h2466b09_1
lzo=2.10=hcfcfb64_1001
markupsafe=3.0.2=pypi_0
matplotlib=3.10.1=pypi_0
menuinst=2.2.0=py312h5da7b33_0
mpmath=1.3.0=pypi_0
multidict=6.4.3=pypi_0
networkx=3.4.2=pypi_0
numpy=2.2.4=pypi_0
openssl=3.5.0=ha4e3fda_0
orjson=3.10.17=pypi_0
packaging=24.1=py312haa95532_0
pango=1.56.3=h0c53d3b_1
pcre2=10.44=h3d7b363_2
pillow=11.2.1=pypi_0
pip=24.2=py312haa95532_0
pixman=0.44.2=had0cd8c_0
platformdirs=3.10.0=py312haa95532_0
pluggy=1.5.0=py312haa95532_0
propcache=0.3.1=pypi_0
psutil=7.0.0=pypi_0
pthread-stubs=0.4=h0e40799_1002
pybind11-abi=5=hd3eb1b0_0
pycosat=0.6.6=py312h2bbff1b_1
pycparser=2.21=pyhd3eb1b0_0
pydot=3.0.4=pypi_0
pygraphviz=1.14=py312h4c6beb1_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=py312haa95532_0
python=3.12.9=h3f84c4b_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.12=6_cp312
reproc=14.2.4=hd77b12b_2
reproc-cpp=14.2.4=hd77b12b_2
requests=2.32.3=py312haa95532_1
ruamel.yaml=0.18.6=py312h827c3e9_0
ruamel.yaml.clib=0.2.8=py312h827c3e9_0
scipy=1.15.2=pypi_0
setuptools=75.1.0=py312haa95532_0
six=1.17.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tensordict=0.8.0=pypi_0
tk=8.6.13=h5226925_1
torch=2.5.0=pypi_0
torch-geometric=2.6.1=pypi_0
torchrl=0.6.0=pypi_0
tqdm=4.66.5=py312hfc267ef_0
truststore=0.8.0=py312haa95532_0
typing-extensions=4.13.2=pypi_0
tzdata=2024b=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_1
urllib3=2.2.3=py312haa95532_0
vc=14.40=haa95532_2
vc14_runtime=14.42.34438=hfd919c2_26
vs2015_runtime=14.42.34438=h7142326_26
wheel=0.44.0=py312haa95532_0
win_inet_pton=1.1.0=py312haa95532_0
xorg-libice=1.1.2=h0e40799_0
xorg-libsm=1.2.6=h0e40799_0
xorg-libx11=1.8.12=hf48077a_0
xorg-libxau=1.0.12=h0e40799_0
xorg-libxdmcp=1.1.5=h0e40799_0
xorg-libxext=1.3.6=h0e40799_0
xorg-libxpm=3.5.17=h0e40799_1
xorg-libxt=1.3.1=h0e40799_0
xz=5.4.6=h8cc25b3_1
yaml-cpp=0.8.0=hd77b12b_1
yarl=1.20.0=pypi_0
zipp=3.21.0=pypi_0
zlib=1.3.1=h2466b09_2
zstandard=0.23.0=py312h4fc1ca9_1
zstd=1.5.6=h0ea2cb4_0
